Fix animation first frame where root layout was reset in release but not in debug builds
Summary: In debug builds we are not releasing root node of LayoutState after constructing it, while in release builds we do. This affected root bounds animation as we were reading the transition key of the root from root node. While it worked on debug, it didn't work on release for the reason above. Initializing root transition key and using it separate from root node fixes the issue.

Reviewed By: IanChilds

Differential Revision: D7995230

fbshipit-source-id: 4099c83b0dbd82b671b79ac7825e2b5cb1c983f6