RenderThread AnimatedRoundedRect drawable
Summary:
this creates a drawable that can draw a rounded rect and animate properties on it using the render thread. AnimatedRoundedRectDrawable delegates to the appropriate implementation of an AnimatedRoundedRect depending on whether the device supports display lists or not.
The canUseRenderThread check is doing the same checks that we do in DisplayList#createDisplayList

More context on the api: https://fb.quip.com/VRm8AD5V5i67

Depends on D7876733

Reviewed By: astreet

Differential Revision: D7877392

fbshipit-source-id: 425bf22f142bd871bd36e6d9e72a7be18b3c2f38