Replace SimpleArrayMap with synchronizedMap in TreeProps and RenderInfo
Summary: `ArrayMap` implementation is not thread-safe and we have lots of crashes when accessing it from multiple threads. TreeProps and RenderInfo which use ArrayMap can be accessed from multiple threads so similar crashes are happening there. I'm replacing it with synchronized hashmap, which is slower and require more memory than ArrayMap, but will work correctly. I'm open to suggestions if you know better data structure than that.

Reviewed By: IanChilds

Differential Revision: D8004186

fbshipit-source-id: d3bdeddee7f2918959682329b3f2486079d5b6ba