Restructure Circle CI config.yml to use Workflows
Summary:
This PR leverages CircleCI Workflows in order to build samples & run tests in parallel.

It first builds the whole project and then starts to run the following jobs in parallel:

1) `sample` build with BUCK
2) `sample-barebones` build with BUCK
3) `sample-codelab` build with BUCK
4) `sample-kotlin` build with Gradle
5) `litho-it` tests run with BUCK
6) `litho-it-powermock` tests run with BUCK
7) Tests run with Gradle

Every job that runs tests, also stores test results and artifacts. Note on this, the previous `regex` used was not grepping tests results from `litho-it` & `litho-it-powermock`. In order to save those results as well, I modified the jobs that run tests using `BUCK` to output the results in a `*.xml` file respectively.

* Declare any references to be used in following steps
* Declare Jobs that will run in Workflow
  -  Each Job consists of:
     - Container Configuration
     - Workspace attaching
     - Cache restoration
     - Main work execution (checkout code/build/test runs)
     - Store tests results (optional)
     - Store tests artifacts
* Declare the Workflow alongside Jobs execution order and Jobs inter-dependencies

* Currently `Workflow` is based on passy 's Docker image specially created for `Litho`:
1) Based on `circle-ci-android` Docker image
2) Installs `BUCK`
3) Installs `androidSdk`
4) Warms up some caches in order to reduce build time on the first job.

Although this is great, Circle CI caches images on `containers` so it was pretty hard to end up on a container with that specific Docker image cached. It did happen though a few times (luckily enough there were some times I needed quick feedback loop and the container had the image cached), effectively letting the Job setup the environment in 1 second.

By switching to `circle-ci-android` Docker image and pulling inside a configuration job, that would of course run as a first Job in the `Workflow` would drammatically increase our chances to find a cached Docker image and decrease the 5-6 mins of environment setup to 1 second. Considering the number of jobs in the `Workflow` (10 atm) it can help reduce the average completion time at an order of magnitude 😃

* It took me around 3 days or so 🕵️
* Ultimately final working version in my branch on my fork consisted of 59 fixup commits 😂
* It took 269 job runs to finally get it to work 😮
* Average `Workflow` run is 40 mins (will explain below some tips that can improve the time) ⏲
* CI provisioning is a tough task 😂

![Cached image inside container](https://user-images.githubusercontent.com/1070871/39297768-10c17ad8-494d-11e8-8884-a420b9fb8e72.png)

![Final success Workflow run](https://user-images.githubusercontent.com/1070871/39297781-16b11a8e-494d-11e8-9f93-cc8ae944e654.png)
Closes https://github.com/facebook/litho/pull/350

Reviewed By: IanChilds

Differential Revision: D7966601

Pulled By: passy

fbshipit-source-id: 0f7601850d542f32dfae0fc0a1929dcde3dc32c4