Add configuration to wait until layout to compute LayoutState when LithoView has EXACT measure specs
Summary:
This is related to the double measure bug from D7938357: the height case is much harder to detect than width. Our way of tackling that is to postpone computing any LayoutState until layout when the measure spec is EXACTLY/EXACTLY for width/height. We can do this because the measurement of our ComponentTree will not be able to affect our measured dimensions. It also happens that the double measure bug always seems to happen with an exact width and height.

This is theoretically something that can live beside the previous fix since the previous fix handles the bug for width when there's an indeterminate height, but we'll probably end up being fine with just this fix.

Reviewed By: IanChilds

Differential Revision: D8014633

fbshipit-source-id: b327c199d311a3df5ccdc5e8617f4879870df19b