Allow publish_snapshot job to run for master branch only
Summary:
Currently, `publish_snapshot` runs for every commit which adds around 2-3 mins per `Workflow` run but then fails if the commit was not in `master`. Adding a filter on that Job allows it to shut down early instead of spending time restoring not needed caches, only to fail after.
Closes https://github.com/facebook/litho/pull/359

Differential Revision: D8041354

Pulled By: passy

fbshipit-source-id: ec66a67a76945be5efd816ca595fb0ee500aee2f