Keep information about whether ComponentTree has been mounted when component goes out of range
Summary: We rely on whether ComponentTree has been mounted (at least once) to decide whether or not to run appear animation. When component is within computed range in RecyclerBinder this information is kept, however as soon as component goes out of range we lose that information and after it comes back to screen its internal state will tell that it hasn't been mounted before and we will trigger appear animation wrongly. Retaining that state when component goes out of range and applying it when it enters range solves the issue.

Reviewed By: astreet

Differential Revision: D7959274

fbshipit-source-id: fef8b569d02e598012100615bf8324371d994ef6