fix fblite size regression
Summary:
The initial fix (D7939473) for [fb4a e2e tests failing](https://fb.facebook.com/groups/441188112968860/permalink/458541171233554/) was too broad. It pulled in the transitive closure of `LithoView` to every app, which is especially painful for fblite.

This change only forbids renaming. This way, apps that don't need or use LithoView won't include it. But for apps that need it, it won't be renamed. Best of both worlds... hopefully :)

Reviewed By: emma0303

Differential Revision: D8014752

fbshipit-source-id: 60d5b8ae83cab8239abc778a29d69a0165e4450e