Removed AccessibilityRole from Litho Images
Summary:
Anything that has an accessibility role is now being forced to be focusable, due to an update to the ExploreByTouchHelper class in the new support library.  This doesn't fix all of the issues, but it does fix the majority of them, with little consequence and little risk.

The real fix will be to refactor how Litho handles elements with accessibility properties and only delegate to the ExploreByTouchHelper class if an element is already going to be focusable.

Reviewed By: sjkirby

Differential Revision: D8085963

fbshipit-source-id: 8467e7eb9893ed9436ef2124ea74c2f32e07a1b3