Add integration test for Section tree props
Summary:
Testing TreeProps in Sections with just unit tests is not possible without
stubbing out massive parts of the runtime which ultimately influences its
behavior. In order to get more confidence for upcoming changes, I want to have
an end-to-end test which involves a RecyclerView and a GroupSectionSpec.

Right now, it documents the old behavior, which makes it easier to observe the
effect of the upcoming change. That change will enable Sections to propagate
TreeProps through Sections and ultimately allow Sections to manipulate treeprops
themselves.

This sets up a new high-level project as we need a full activity to run this and
want to have very fine-grained dependencies for this. It also surfaced some
setup issues with `:litho-espresso` that had gone unnoticed before.

Reviewed By: IanChilds

Differential Revision: D8075296

fbshipit-source-id: c01748274b792b14d2c89be3a8b0c825b6f3284c