Separate flags for view and layout
Summary:
We bundled all these flags into a single int, but we need to separate them, since the view flags shouldn't be updated every time (next diff), but the layout ones should.

I had to make a number of methods non-static but to be honest this is quite a bit nicer as the user no longer has to pass in the correct value.

Reviewed By: marco-cova, mihaelao

Differential Revision: D8088667

fbshipit-source-id: 8c249c9e1e83b5d43fb7420ff677c967c01c753a