Don't reset View flags in MountItem when we update the layout output
Summary:
With the previous code, we could run into the following problem:
- Set up a MountItem with mContent being a View that is not clickable, so save that mContent is not clickable.
- Set a click handler on the mContent when it is mounted, mContent is now clickable.
- Call init on the MountItem - this will check the mContent again to see if it is clickable, and now it is, so we mark it as clickable.
- When we are unmounting mContent we think that it was always clickable so leave it in that state, when actually originally it was not.

This class was a bit messy, because we were calling `init()` when really we were just updating. I've separated it into `init()` and `update()` - you shouldn't now call `init()` on anything that was not freshly acquired, and `update()` will update layout flags (which can change with the new output), but not view flags (which don't change as the content hasn't changed).

Reviewed By: marco-cova

Differential Revision: D8089294

fbshipit-source-id: 35f8d4b0f76616280850fa4e7891a1be084c59f5