Rework docs toggler
Summary:
The old toggler code was really messy and relied on rebuilding incorrect HTML on
the fly to make the buttons at the of the getting started page clickable. It
also required some global singletons and mutating some global state to configure
stuff.

This is now much simpler and just looks for a certain HTML structure, making it
interactive. It also works just on top of URL hashes, so they can be shared more
easily. The old script had the *idea* of finding IDs in the hierarchy and
enabling the corresponding tab, but in practice that never worked.

This should make it easier going forward to add more tabs like this, which is
the plan to add more Kotlin examples.

Reviewed By: marco-cova

Differential Revision: D8117374

fbshipit-source-id: 0ba0e80cb2ea3f64bea4b5a735c6a1199049c1f2