Animate arbitrary view with render thread
Summary:
Delegate ViewProperty animations to either the render thread or to view.animate. This enables view animations on the render thread.
This doesn't yet handle paint, paint alpha and paint stroke animations since they are not supported on view.animate.

I'll add support to animate these properties on a drawable in a separate diff; I can make them work for render thread by setting the drawable's canvas as a target, but not sure how to animate these properties on drawables using main thread animations. astreet do you know if there's a straightforward way to achieve that?

Reviewed By: astreet

Differential Revision: D8024850

fbshipit-source-id: be1d0fd9af7bd49d1b3360d8c382c077ef6a94f3