Treat all UNSPECIFIED measure specs as compatible
Summary:
Right now we check if the specs are exactly the same. Unfortunately, UNSPECIFIED specs have an ignored size part still, and that size part, while normally 0, can also be an arbitrary number.

This diff explicitly changes the compatibility logic to not rely on UNSPECIFIED specs being exactly the same.

Reviewed By: astreet

Differential Revision: D8132072

fbshipit-source-id: a81cd510c13735b9bcf8ca40a7dbb0dbebb52217