Extract Common Prop Defaults and put them in SpecModels
Summary:
We want to be able to set defaults for a common prop - for example, a Button might want to default to accessibility role of BUTTON, or we might want to make Text always important for accessibility. This adds the basic framework for getting Common Prop Defaults into a SpecModel - the next diff will actually use this stuff.

As you can see, it's a bit unwieldy adding stuff to the spec models these days, but hopefully it's quite simple.

Reviewed By: marco-cova

Differential Revision: D8182319

fbshipit-source-id: da8c60c61e721bbc5b026db3a76f884e68aa6179