Ensure correct order for mounting content with the same bounds
Summary:
Consider a Component made up for two Components A and B that have the same bounds (e.g using absolute positioning). We want A to always be added before B (that is our expected drawing order and clients might make other assumptions based on that). Right now, we store the bottoms and tops of our Components so that we can make incremental mount more efficient. We distinguished between hosts and non-hosts for the order of the bottoms and tops, but not drawing order.

Therefore right now the bottoms would be in order "A, B, root", which would mean that when scrolling up, root would be added first, then B, then A. In actual fact we want the order to be "B, A, root". We can achieve this by changing LayoutOutput to take in an index, and then using that index in the Comparator. We can also simplify the comparator quite a lot by using the fact that we can return any positive or negative number, it doesn't have to be 1 and -1.

Reviewed By: astreet

Differential Revision: D8222867

fbshipit-source-id: 5a6bc8fe9406b5a0ecea2a0a0f207f03976276fa