Support Section TTI perf logging for items shown on the screen
Summary:
Since Android doesn't provide callback for us when a view is drawn complete, we use fbsystrace to trace the draw behaviour and decide to override the `RecyclerView#drawChild()` to notify a RecyclerBinder that a item has been rendered.

To prevent hurting the perf, the notification is only triggered when the child has registered the RenderCompleteEvent handler and its renderState isn't RENDER_DRAWN yet.

The RenderCompleteEvent is dispatched to section asynchronously. The event includes two arguments, hasMounted means if the item is on the screen or not, and timestampNanos is the timestamp when render is complete, the unit is nanosecond.

Reviewed By: astreet

Differential Revision: D7813672

fbshipit-source-id: f2a49486e6faf58602f36d56c06687804eccda1f