Move onCreateMountContent to have Object param
Summary: We want to provide onCreateMountContent with an Android context, not a ComponentContext, as ComponentContext is a heavier object that holds onto things that should not be stored in our pools. This allows us to have both Context and ComponentContext accepted as a param as an immediate stage while we convert (not expected to be more than a day or two, it's really just to allow us to land with drop-conflicts enabled.

Reviewed By: astreet

Differential Revision: D8202688

fbshipit-source-id: 99ac3e5785630860d71e01f570af91d4a302c66d