Throw if setting setAllowRunningAsynchronously to false on the RenderNodeAnimator
Summary:
AnimatorSet will call setAllowRunningAsynchronously(false) on the animators it's playing before starting, so even if we're using RenderNodeAnimator we'll still be animating on the main thread. This adds a RenderNodeAnimatorAsync(for lack of a better name) that overrides setAllowRunningAsynchronously to crash if it's turning off async animations. Next step is to create an AnimatorSet implementation that plays well with RT
Depends on D8024850

Reviewed By: muraziz

Differential Revision: D8160338

fbshipit-source-id: c36b1d41bafdb9b157e4148001c049bb0b20d3f0