Remove TestComponentContextWithView
Summary: For some reason (I guess there used to be restrictions before we joined Component/ComponentLifecycle and Component/ComponentLayout) we were using a pretty weird technique to get a View from a TestViewComponent, where we'd set the View to the Context and then find the view from that context. Instead, we can just set it directly on the TestViewComponent.

Reviewed By: astreet

Differential Revision: D8247218

fbshipit-source-id: 9593af079dd2008aec51116ff17258d0b8247a57