Add isEquivalentTo to EventHandler
Summary: When we do `shouldUpdate()` for a Component that has an EventHandler as Props, we'll always return true, because the comparison of the event handlers will find that they are different, since we're creating a new one each time. Instead we need the ability to compare them, which this provides. I'll integrate that into the next diff.

Reviewed By: passy

Differential Revision: D8236784

fbshipit-source-id: 9d589b28d7ab111cf8c5b884088665d569115158