Preserve TreeProps through SectionContext
Summary:
Ensure that the `TreeProps` stored in the `ComponentContext` used to create a `SectionContext` is copied over. This needs to happen both when constructing a new context and in the RecyclerBinder.

This does *not* address the problem of modifying TreeProps within a Section yet, as can be seen in the integration test. Otherwise, we would expect the end result to contain a "root:section:leaf" tag.

Reviewed By: IanChilds

Differential Revision: D8077060

fbshipit-source-id: b4c5535566e19813aca72a166adcf5f94978a2db