Port `animations` to `sample-kotlin`
Summary:
This PR follows a similar path as the previous one :) It ports `animations` package from Java `sample` to Kotlin `sample-kotlin`.

A nice side effect of it is that even more parts of the framework are being "tested" in Kotlin code, which will allow catching Kotlin-related issues earlier.
Closes https://github.com/facebook/litho/pull/379

Reviewed By: muraziz

Differential Revision: D8235929

Pulled By: passy

fbshipit-source-id: 420500d34e04fa5f078670bdf22f99e595ae9926