Set up ErrorBoundary auto-wrapping
Summary:
Add a configuration option to wrap every component app-wide in a root component, which can be used to set up a catch-all error boundary for every component tree.

Because of the way `.config` exists as a leaf module, we cannot add either the configuration option or the necessary callback to the module as it would create the circle `core -> config -> core`. The current setup is only temporary as this will ultimately move to an opt-out `ComponentTree` option with a barebones error boundary that we can ship as part of core, but we need some more data first and want to add support for additional delegate methods.

Reviewed By: emilsjolander

Differential Revision: D8237330

fbshipit-source-id: 1f4c0a96e1b81f1d25a1cade83913bbd7b224c88