Fix error boundary wrapping in ComponentTree
Summary:
The previous version would unconditionally wrap all roots that are set in the
ComponentTree in an ErrorBoundary. This would also happen for calls to setRoot
coming from measure or updateState which re-use the previous root (or rather a
shallow copy of it). That would cause the error boundary to be wrapped in
another error boundary. Once the error boundary tree gets too high, we'd easily
run into StackOverflow issues in all sorts of contexts.

This diff splits those method calls that reuse a component tree from the others
that set a new root to prevent that.

Reviewed By: mihaelao

Differential Revision: D8312558

fbshipit-source-id: 1766bc1510ab1ffc3bd0fafac2ad47b58dfa3dac