Force a mount when our visible rect has changed
Summary:
The example of a case where this is required is when we show a keyboard on top of a view, so when it is mounted, we don't mount the content behind the keyboard as it is not visible. Then when the keyboard is removed, we don't bother mounting content as the view bounds are the same.

I don't know of a better way to fix this that would catch this case, but it is a shame to have to make this change - previously we were just comparing rects, now we're going up the hierarchy. Still, performLayout should not be called *too* often, so I think this should be okay.

I have put this behind my GK for visible bounds as it was only reported in this case and it would be nice to have some way to turn it off.

Reviewed By: pasqualeanatriello

Differential Revision: D8314597

fbshipit-source-id: 33b828fae3aac630a0a46f1f1db5a074d3b92a1d