Move listeners from bind to mount in VerticalScroll
Summary: We don't need to do this every time we bind/unbind, we can do it in mount. To do this we move the code to LithoScrollView - we track scrolling in onScroll and still use the OnPreDrawListener (I tried moving that to dispatchDraw, but it seemed to be too late as it ended up flickering).

Reviewed By: astreet

Differential Revision: D8237315

fbshipit-source-id: 0434f2712e89418e0fd838aebbfaf1b6ffcbe80e