Use new component instance when layout is created from multiple threads
Summary:
Addresses https://github.com/facebook/litho/issues/360
When a component is passed as a prop we always make a shallow copy before creating layout. This is handled in the generated code for the builder method. It's not so trivial to handle the case when the Component is passes a prop indirectly, either as an item in a list or wrapped in another data structure.
This checks if layout has been created for a component from a different thread, in which case we need to move forward in the current layout creation with a copy of that component.

Reviewed By: pasqualeanatriello

Differential Revision: D8281975

fbshipit-source-id: 122f47e92b713eeeffeebddbfdae2d50d02465d5