RecyclerBinder fill viewport on multiple threads
Summary:
compute layouts to fill the viewport in parallel on threads from a pool.
Working on unit tests, but sending this out to get feedback on the approach.

The main thread schedules a task for each thread in the pool, then waits for them to complete. We wait for the layouts to be completed in order, then when the next expected task completes we check if we have a full viewport, then unblock the main thread if that's the case. We'll at most be blocking the main thread unnecessarily during the last scheduled batch if the item that's needed to fill the viewport gets completed last, before the items that are not in the visible viewport.

Reviewed By: astreet

Differential Revision: D8314402

fbshipit-source-id: fd9c9667c7248c02238c35a0e3f8ea8bc9461ee6