Ship lazy layout for exact measurements
Summary:
This is part of the fix for android's double measure bug. It postpones ComponentLayout calculation from measure to layout is the measured dimensions are exact: this means that if we get remeasured with different exact measurements, we don't compute two layouts.

Deltoid results: https://fburl.com/deltoid3/ylrbxcpu (big-ish improvements for watch tab, small improvements for groups mall)

Reviewed By: IanChilds

Differential Revision: D8434171

fbshipit-source-id: 85b5c9aa2abd960d264071d6d358ecbbb6e50c78