Skip logging LithoView zero height when view is being removed in prelayout
Summary: We have been logging cases where LithoView is measured with 0 height in vertical list. From the logs that we got there are bunch of cases where 0-height is not affecting users like when item is being removed and predictive animation in RecyclerView tries to layout removing item in pre-layout step. That case is not critical (does not affect users) and filtering those cases out will improve the quality of signal we get from logging.

Reviewed By: astreet

Differential Revision: D8380115

fbshipit-source-id: c440e6fe19dd27f356eeee51b21e3ed7f6693005