Move onSetRootAndSizeSpec outside of synchronized block to prevent deadlock
Summary:
We received a report that there's a deadlock in ComponentTree and ComponentTreeHolder, P59711548 shows the call stack when deadlock is occurred:

Thread 1 is running `ComponentTreeHolder.checkWorkingRangeAndDispatch()`, it has the lock on the ComponentTreeHolder object, and it's calling `ComponentTree.checkWorkingRangeAndDispatch()`, so it is blocked waiting the lock on ComponentTree.

Meanwhile, thread 40 is inside `ComponentTree.calculateLayout()`, so it has the lock on ComponentTree, and is calling `mMeasureListener.onSetRootAndSizeSpec()`, which is calling into `ComponentTreeHolder.getMeasuredHeight()`, therefore blocked waiting for the lock on ComponentTreeHolder.

I think `mMeasureListener.onSetRootAndSizeSpec()` don't need the lock on ComponentTree, move it outside out synchronized block should fix the deadlock problem.

Reviewed By: astreet

Differential Revision: D8510184

fbshipit-source-id: 39499179066dc2c35567bde2a2c9cf08548eb3f0