Fix layout diffing when layout state is calculated twice before mount
Summary:
When we're doing layout diffing, we would just check against the last Component id to verify whether or not we should use the shouldUpdate values from the layout diffing for our mount diffing. However, that can lead to a bug where we don't mount if we set a root on a ComponentTree twice before we mounted it - the second time that we set the root we will think that nothing needs updating, and that's what we'll tell the MountState, so we don't update when we should.

Instead of just comparing to the ComponentTree, we should compare to the last mounted layout state. To do this we add ids to LayoutState and we use that.

Reviewed By: marco-cova

Differential Revision: D8531551

fbshipit-source-id: 040fa499856b3e2989efe780f4246d2bd852c9ef