Use LayoutManagerOverrideParams in VerticalScroll
Summary:
We need this change in order to make things work with fillViewport (without it the ScrollView tries to remeasure internal content with wrap content so it doesn't end up filling the viewport), which I'll add in the next diff, but I wanted to add this first to make it more obvious what the change does.

We record the height and width of the child content in onBoundsDefined, and from there we can simply pass that size to the LithoView that we mount within the scroll view. That LithoView sets its own layout params with the height and width that it wants.

Reviewed By: astreet

Differential Revision: D8488194

fbshipit-source-id: 29552642e7d468f1d987ae5f82f5ab275083a73d