Updates Litho Events Overview
Summary:
This documentation currently shows creating an event handler via a Prop, but this seems like a strange thing to do -- Litho automatically creates a prop named after your event when you specify events in LayoutSpec. If you follow this example:

1) Your code will not compile if you give your Prop the logical name (fooEventHandler).
2) Otherwise you end up with two event handler props with different names

I think we should suggest that people use getFooEventHandler() instead. This is the common practice in our code.

Reviewed By: IanChilds

Differential Revision: D8430746

fbshipit-source-id: 3749e7f3192e27bb4cdfe405521570338add8496