Experiment to always perform inc mount in layout
Summary: Tried something similar in D8314597, but that caused some bugs that I couldn't repro and therefore couldn't investigate. This time the logic is simpler - just always perform incremental mount if it is turned on and we didn't do it in the layout call. This shouldn't have too much overhead when there is nothing to do, and I've put it behind a GK so that we can turn it off easily if we need to.

Reviewed By: pasqualeanatriello

Differential Revision: D8540915

fbshipit-source-id: 1964a7fba67c6c9653420c38dc01318c643833b2