EditText should not reset text if you specify "null"
Summary:
`EditText` was continually resetting the view's text to null if you explicitly specified `null` for the prop.

Normally you would work around this with the `stateUpdatePolicy` prop, but that has some bugs with asynchronous layout that are not yet resolved.

In the interim, change the API contract so that if you explicitly specify null, the text field is simply left untouched. If you want to explicitly reset the field's text, use the empty string.

Reviewed By: astreet

Differential Revision: D8590956

fbshipit-source-id: a160abf75cd10e5c49969de1fc085bbdaa607504