Add initialText prop to EditTextSpec
Summary:
What if you want to configure the initial text on a Litho EditText, but you then want Litho to subsequently take a hands-off approach and not change the EditText's text property?

Previously this was impossible. Now you can do it with `initialText`.

This really calls out for a deeper redesign of `EditText`, which is on the way, but this is a start.

Reviewed By: gmoeck

Differential Revision: D8595780

fbshipit-source-id: 1347f249d3b73bab5e2a8d98a684c554a53b92cc