Move common props to a better place
Summary: Right now, the framework for CommonProp is pretty bloated, and it doesn't tie in nicely with Prop. Instead, we can just make it a param on Prop and integrate into the framework for Prop, by calling `super.(...)` if it is a common prop. In future, we can also add the ability to not call `super` so that you can "override" a common prop, but for now we want this to do the same thing as setting up your own CommonProp.

Reviewed By: astreet

Differential Revision: D8614898

fbshipit-source-id: ce96c764fb9477d50e083220606c8af9f81216c2