Copy list when used as a default
Summary: When we're using var args we add a new value to the list if the list already exists. If we have a list specified as the static default, then that list was always being updated, which is wrong. Instead, let's just make a copy of the list when we set the default value on the prop.

Reviewed By: passy

Differential Revision: D8658778

fbshipit-source-id: 77d3ad70d747eed4f119aea98a24ac5b8526b06e