Only makeShallowCopy of Component in Builder
Summary: We don't care about making copies of Sections in the Builder, because after the previous diff we can just do that when we get all the children. This code was also buggy because it depended upon the type of the Spec as to whether we would make a shallow copy or not. Now, we always do for props that are Components, but not for props that are Sections.

Reviewed By: passy

Differential Revision: D8713052

fbshipit-source-id: 591dccc60f91b3bdfdcaf15fe4e26079cdf602f9