Don't re-collect animations on double measure
Summary:
This is a fix for a bug when combining UpdateStateWithTransition, bounds animations, and a double measurement!

We collect transitions early (in onMeasure) when we have bounds animations so that we can provide the correct measurement for the start of the animation.

Because of the specific container layout, we're having onMeasure called twice in the same traversal (this is not great, but is not incorrect behavior on android -- sometimes you really have to measure twice). However, this second measure was causing us to collect transitions again and wipe out the set of transitions where we'd consumed the one from OnUpdateStateWithTransition, thus losing that transition.

Reviewed By: muraziz

Differential Revision: D8732261

fbshipit-source-id: c11f6ea53eac26b844d2102d5614e87cd783fc3a