Add SpinnerSpec to litho widgets
Summary:
This diff adds a `Spinner` component that is derived from Androids `SpinnerView`.

Since this component is not mounting the SpinnerView and emulating it, it will have the same appearance no matter which API level it's rendered on.

It's slightly less general than the SpinnerView in that it only accepts a list of options and the ability to customize the view.
However, if the need arises we can generalize it further and use a `ListItem` pattern or similar to bind custom views/components with custom data.

However, such cases are very rare looking at my experience with Android.

**Caveats**
This view is actually close to the one defined in the material design guidelines that the one shipped with the platform in the following ways:

* The caret is rotated when the dropdown is show
* The drop down does not cover the currently selected value

**Material design ref:**
https://material.io/design/components/menus.html#exposed-dropdown-menu

Reviewed By: IanChilds

Differential Revision: D8670252

fbshipit-source-id: d5ea373d4d5de112bd75fa2dff6a65bfe11e6177