Implements text truncation with custom ellipsis in litho TextSpec
Summary:
As prototyped in D7541329, this enables clients to pass a 'customEllipsisText' prop which provides a custom CharSequence to show at the end of truncated lines.

One of the biggest requests from FDS design has been the ability to truncate text on Android with a 'See More' label at the end of the last line of content. Currently, most text truncation in fb4a is handled by truncating after a fixed number of characters. I went ahead and added an example use case for this feature to our code to validate the approach, but I'm happy to move that to a follow-up diff if desired.

Reviewed By: astreet

Differential Revision: D7765570

fbshipit-source-id: 636d1686ff3b4eeff75e7729a84e1dd8f16d8405