More efficient fillViewport for VerticalScroll
Summary:
(hat-tip to astreet for the idea)
Previously we were measuring once with unspecified height, and then if we were fillViewport we would remeasure with EXACT height. Instead, we can just set the minHeight of the component that we are passed.

Because we are doing the wrapping inside the Spec, if we get passed the same component we will not early return, because the wrapper will be a different component. However, that seems to be a pretty rare case that we shouldn't worry about in return for this gain (and is only true for fillViewport anyway).

Reviewed By: astreet

Differential Revision: D8783013

fbshipit-source-id: 408f95ca2a9d5b4247ef501288be9b62e80fb2a4