Fix visibility events for items that are removed from the tree
Summary:
You can see the case that this is fixing in the test case that I've added - we have a component with a visible handler that is then completely removed from the mount state. Previously we would not do anything in this case, because when we iterate through the visibility outputs on the new layout state we would not find that visible handler anywhere. Instead, we need to look at the visibility items that we have stored that we do not account for in the loop of visibility outputs - if we do not account for a visibility item then it is no longer in the tree at all, and therefore we should send invisible and focused events and remove it from the array of visibility items that we have stored.

I couldn't think of a nicer way to write this code, if you don't like having to store that bit of state then I am open to ideas :)

Reviewed By: marco-cova

Differential Revision: D8800110

fbshipit-source-id: 2096be443ec615195ffcd5d5e7294ca7f0ea4cea