Ignore uninitialized LayoutParams from the RecyclerBinder (#395)
Summary:
It's possible that RecyclerBinder doesn't have its size set by the time the RecyclerView measures itself (and lays out its children within its measure call). Due to the bitmasking on the measurespec int, the UNINITIALIZED -1 value that is sent in results in the cell being measured with width 0. Since the MeasureSpecs from the View's native onMeasure are still valid, we can just ignore the UNINITIALIZED value from the OverrideParams.
Pull Request resolved: https://github.com/facebook/litho/pull/395

Reviewed By: IanChilds

Differential Revision: D8810035

Pulled By: passy

fbshipit-source-id: f40713425b3095e12b4e49ceb180e8102c2e7d1a