Allow more than one visibility event per InternalNode
Summary: When we try to add a visibility event to an InternalNode that already has a visibility event of that type, then we will create a DelegatingEventHandler that will dispatch events to both of the handlers when a visibility event occurs. This means that we only pay extra memory cost in the rare instances that we have more than one visibility handler.

Reviewed By: astreet

Differential Revision: D8952417

fbshipit-source-id: be18f18489600f56c0501b10381c70e0201f02b3