Replace SimpleArrayMap with LinkedHashMap for transitionKey to output mapping in LayoutState
Summary:
`SimpleArrayMap`/`ArrayMap` is not safe to use in multithreaded environment as it is not designed to do so in the first place. What's more interesting is that `SimpleArrayMap` uses static internal cache for its backing array and once it gets corrupted in one usage in multithreaded environment, it affects *ALL* other current usages even if some of other usages are legit (accessing it only from single thread). `SimpleArrayMap` has `freeArrays` and `allocArrays` methods which are self-explanatory and let's say multi-threaded usage puts array into the cache which gets corrupted, and some other legit case (like the one in the diff, `mTransitionKeyMapping` is used only in UI thread) next pulls that cache it crashes the app (Someone did an excellent analysis on that: https://gist.github.com/LanderlYoung/579872afffc62a5f837e654a6f1eab89).

Apparently if you have at least one usage of `SimpleArrayMap` in multi-threaded environment, all other usages within same process will become at risk of crashing with `ClassCastException`s. Only correct way of using `SimpleArrayMap` is to *guarantee* that there is no usage of it in multi-threaded environment. While it is nearly impossible to guarantee that with big apps, safest way is not to use it at all.

Reviewed By: sergeynv

Differential Revision: D8966285

fbshipit-source-id: e59eec292270d760e090882c34d7a8c6fed6c664