Make CommonProps accessible from a Component
Summary: There are times when you might want to know what CommonProp was set on a Component. For example, if you wrap a Component and add a click handler, it will override the clickHandler that was set on the delegate. You might instead want to get the clickHandler from the delegate and fire that event as well.

Reviewed By: passy

Differential Revision: D8966415

fbshipit-source-id: 1e2658a22577c1efc764c08c617831e0d6fd7eab